Only automatically scale ImGui size if it doesn't already do that itself
ImGui scales everything by framebuffer_size / window_size
(window size is in logical points that are not the same as pixels on
 High-DPI screens, at least on some platforms).
so if framebuffersize / window_size == content_scale
(from glfwGetWindowContentScale) then we shouldn't scale the font
any further.
if however the content_scale is bigger than that ratio, for example on
KDE/X11 with window-scaling enabled, the font is scaled up accordingly
after all.