Select KTX2/Basisu transcoding target based on driver support
probably ASTC gives the best quality, as UASTC is a subset of that,
so if the GPU/driver supports that, it should be used.
If it doesn't, use BC7, if that is supported.
If neither is supported, use uncompressed 32bit-per-pixel RGBA

also removed some commented out code and fixed Mac compiler warnings